Rename package from errors -> sentry
